+++ libraries/tools/kotlin-stdlib-gen/src/templates/Generators.kt
@@ -944,7 +944,7 @@ object Generators : TemplateGroupBase() {
         }
         body(Sequences) {
             """
-            return buildSequence result@ {
+            return defineSequence result@ {
                 val iterator = iterator()
                 if (!iterator.hasNext()) return@result
                 var current = iterator.next()