+++ libraries/tools/kotlinp/test/org/jetbrains/kotlin/kotlinp/test/KotlinpCompilerTestDataTest.kt
@@ -12,7 +12,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import java.io.File
-import kotlin.coroutines.experimental.buildSequence
+import kotlin.coroutines.experimental.defineSequence
 
 @RunWith(Parameterized::class)
 class KotlinpCompilerTestDataTest(private val file: File) {
@@ -41,7 +41,7 @@ class KotlinpCompilerTestDataTest(private val file: File) {
                 "compiler/testData/serialization/builtinsSerializer"
             )
 
-            return buildSequence<Array<*>> {
+            return defineSequence<Array<*>> {
                 for (baseDir in baseDirs) {
                     for (file in File(baseDir).walkTopDown()) {
                         if (file.extension == "kt") {